abstract function for loading

abstract function for loading

Usage

getAbstract(
  ident,
  from = pwd(),
  addAsLink = TRUE,
  addIdToName = TRUE,
  caption = "",
  description = "",
  folderName,
  func,
  ...
)

Arguments

ident

path, resource or entity ID of the picture

from

used for relative paths, by default pwd is used, which is initiated with the step that started improveR

addAsLink

creates a link in inventory if TRUE, warn improveClean has to be run at the end

addIdToName

adds the entity ID to the filename, enity version ID if a specific version is retrieved

caption

by default entityID and lastmodified are the caption, here alternative text can be provided

description

by default filename is the caption, here alternative text can be provided

folderName

where to store the file

func

the processing function for this type

...

pass over arguments for the read function