loadResourceVersionFromServer

Loads a resource version directly from the server by the entity version ID. The results are returned as a data frame or a list of data frames. The dates are also converted to POSIX dates via convertImproveTimestampToPosix. entityVersionId can be a list.

Usage

loadResourceVersionFromServer(entityVersionId, invalidatesReproducibility = T)

Arguments

entityVersionId

the entity version id of the resource

invalidatesReproducibility

this flag may only be changed by internal functions.

See also