setProcessVariables

setProcessVariables

Usage

setProcessVariables(
  stepId,
  processId,
  processType = "main",
  name = "Main",
  selected = TRUE,
  runserverId = "",
  toolId = "",
  runserverToolId = "",
  toolIgnorePatterns = "",
  mainProcess = T,
  gridTool,
  toolArguments = NULL,
  position = 1,
  parentProcessId = NULL,
  toolBrowserUrl = NULL,
  toolDeletePatterns = NULL,
  toolStreamablePatterns = ""
)

Arguments

stepId

id of the step the variables refer to

processId

id of the process the variables refer to

processType

processType

name

name

selected

selected

runserverId

runserverId

toolId

toolId

runserverToolId

runserverToolId

toolIgnorePatterns

toolIgnorePatterns

mainProcess

if this process is status relevant, there needs to be exactly one mainProcess

gridTool

gridTool

toolArguments

toolArguments

position

the position in the process list

parentProcessId

only to use when processType equals sub

toolBrowserUrl

a URL improve uses to automatically open while the step is running

toolDeletePatterns

files that wont get checked in

toolStreamablePatterns

file that can be streamed to monitor the step